+++ src/mjlab/tasks/velocity/config/go1/__init__.py
@@ -5,7 +5,7 @@ from .env_cfgs import (
   unitree_go1_flat_env_cfg,
   unitree_go1_rough_env_cfg,
 )
-from .rl_cfg import unitree_go1_ppo_runner_cfg
+from .rl_cfg import unitree_go1_flat_ppo_runner_cfg, unitree_go1_ppo_runner_cfg
 
 register_mjlab_task(
   task_id="Mjlab-Velocity-Rough-Unitree-Go1",
@@ -19,6 +19,6 @@ register_mjlab_task(
   task_id="Mjlab-Velocity-Flat-Unitree-Go1",
   env_cfg=unitree_go1_flat_env_cfg(),
   play_env_cfg=unitree_go1_flat_env_cfg(play=True),
-  rl_cfg=unitree_go1_ppo_runner_cfg(),
+  rl_cfg=unitree_go1_flat_ppo_runner_cfg(),
   runner_cls=VelocityOnPolicyRunner,
 )